+++ src/methods/$last/$last.js
@@ -1,0 +1,7 @@
+import { $lastOr } from '../$last-or/$last-or';
+
+export function $last(iterable) {
+  return $lastOr(iterable, undefined);
+}
+
+export default $last;